Grinch: removes Christmas Joy feature
The option was already commented out in the functions.php file. I just
took it a bit further and deleted the styles, function, hooks, and
other stuff left over from the seemingly defunct Christmas Joy feature.

This change was made on a topic branch named "grinch". If this commit
has been merged into the master branch, that means it was fully
discussed and approved by Andrei.