Use admin_url() to eliminate relative redirects
Technically, relative URLs should not be used in Location headers.
In fact, doing so can cause problems; WordPress core ticket #14062
(http://core.trac.wordpress.org/ticket/14062), in fact, resulted
in a fix to that problem for WP's code.

So, this change wraps relative URLs used in functions.php redirects
with the admin_url() function, which returns an absolute URL.